admin_form/requirements.txt
flask
supabase==2.5.0
python-dotenv